+++ internal/services/controllers/task/controller.go
@@ -1065,7 +1065,6 @@ func (tc *TasksControllerImpl) handleProcessUserEvents(ctx context.Context, tena
 
 // handleProcessEventTrigger is responsible for inserting tasks into the database based on event triggers.
 func (tc *TasksControllerImpl) handleProcessUserEventTrigger(ctx context.Context, tenantId uuid.UUID, msgs []*tasktypes.UserEventTaskPayload) error {
-	opts := make([]v1.EventTriggerOpts, 0, len(msgs))
 	eventIdToOpts := make(map[uuid.UUID]v1.EventTriggerOpts)
 
 	for _, msg := range msgs {
@@ -1083,8 +1082,6 @@ func (tc *TasksControllerImpl) handleProcessUserEventTrigger(ctx context.Context
 			Scope:                 msg.EventScope,
 			TriggeringWebhookName: msg.TriggeringWebhookName,
 		}
-
-		opts = append(opts, opt)
 
 		eventIdToOpts[msg.EventExternalId] = opt
 	}