+++ sdks/ruby/src/lib/hatchet/clients/rest/lib/hatchet-sdk-rest/models/v1_webhook.rb
@@ -38,6 +38,9 @@ module HatchetSdkRest
     # The static payload to use for the webhook. This is used to send a static payload with the webhook.
     attr_accessor :static_payload
 
+    # Whether to return the triggered event as the response payload when this webhook is triggered
+    attr_accessor :return_event_as_response_payload
+
     class EnumAttributeValidator
       attr_reader :datatype
       attr_reader :allowable_values
@@ -70,7 +73,8 @@ module HatchetSdkRest
         :'event_key_expression' => :'eventKeyExpression',
         :'auth_type' => :'authType',
         :'scope_expression' => :'scopeExpression',
-        :'static_payload' => :'staticPayload'
+        :'static_payload' => :'staticPayload',
+        :'return_event_as_response_payload' => :'returnEventAsResponsePayload'
       }
     end
 
@@ -94,7 +98,8 @@ module HatchetSdkRest
         :'event_key_expression' => :'String',
         :'auth_type' => :'V1WebhookAuthType',
         :'scope_expression' => :'String',
-        :'static_payload' => :'Object'
+        :'static_payload' => :'Object',
+        :'return_event_as_response_payload' => :'Boolean'
       }
     end
 
@@ -162,6 +167,10 @@ module HatchetSdkRest
 
       if attributes.key?(:'static_payload')
         self.static_payload = attributes[:'static_payload']
+      end
+
+      if attributes.key?(:'return_event_as_response_payload')
+        self.return_event_as_response_payload = attributes[:'return_event_as_response_payload']
       end
     end
 
@@ -282,7 +291,8 @@ module HatchetSdkRest
           event_key_expression == o.event_key_expression &&
           auth_type == o.auth_type &&
           scope_expression == o.scope_expression &&
-          static_payload == o.static_payload
+          static_payload == o.static_payload &&
+          return_event_as_response_payload == o.return_event_as_response_payload
     end
 
     # @see the `==` method
@@ -294,7 +304,7 @@ module HatchetSdkRest
     # Calculates hash code according to all attributes.
     # @return [Integer] Hash code
     def hash
-      [metadata, tenant_id, name, source_name, event_key_expression, auth_type, scope_expression, static_payload].hash
+      [metadata, tenant_id, name, source_name, event_key_expression, auth_type, scope_expression, static_payload, return_event_as_response_payload].hash
     end
 
     # Builds the object from hash